feat: support unlocking scroll when modal dialog closes
